refactor: move profile under internal/app
