+++ internal/app/profile/wizard/create_flow.go
@@ -1,15 +1,15 @@
-// File : internal/profile/wizard/create_flow.go
+// File : internal/app/profile/wizard/create_flow.go
 // Deskripsi : Flow wizard untuk pembuatan profile
 // Author : Hadiyatna Muflihun
 // Tanggal : 4 Januari 2026
-// Last Modified : 4 Januari 2026
+// Last Modified : 5 Januari 2026
 
 package wizard
 
 import (
 	"strings"
 
-	"sfDBTools/internal/profile/shared"
+	"sfDBTools/internal/app/profile/shared"
 	"sfDBTools/internal/types"
 	"sfDBTools/pkg/consts"
 	"sfDBTools/pkg/helper"
