+++ internal/app/profile/wizard/select.go
@@ -1,14 +1,14 @@
-// File : internal/profile/wizard/select.go
+// File : internal/app/profile/wizard/select.go
 // Deskripsi : Pemilihan profile secara interaktif
 // Author : Hadiyatna Muflihun
 // Tanggal : 4 Januari 2026
-// Last Modified : 4 Januari 2026
+// Last Modified : 5 Januari 2026
 
 package wizard
 
 import (
 	"fmt"
-	"sfDBTools/internal/profile/shared"
+	"sfDBTools/internal/app/profile/shared"
 	"sfDBTools/pkg/consts"
 	profilehelper "sfDBTools/pkg/helper/profile"
 )
